pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["Christopher David <chris@arcadelabs.co>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10.9"
flask = "^2.3.2"
gunicorn = "^20.1.0"
supabase = "^1.0.3"
python-dotenv = "^1.0.0"
openai = "^0.27.8"

[tool.poetry.scripts]
testum = "api.test:test"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
